refactor: use search params instead of hash to toggle dashboard settings
