+++ dashboard/src/app/dashboards/DashboardCard.tsx
@@ -63,7 +63,7 @@ export default function DashboardCard({ dashboard }: DashboardCardProps) {
             </div>
             <div className='flex w-8 justify-center'>
               <Link
-                href={`/dashboard/${dashboard.id}#settings`}
+                href={`/dashboard/${dashboard.id}?settings=true`}
                 className='hover:bg-muted inline-flex h-8 w-8 cursor-pointer items-center justify-center rounded-md p-0 transition-colors'
                 title='Dashboard Settings'
                 onClick={(e) => e.stopPropagation()}
